[core] Introduce and apply GeoJSONData::create() API
